+++ src/assets/profanity-filter.json
@@ -1,0 +1,6 @@
+[
+    "tea",
+    "cake",
+    "pastries",
+    "pastry"
+]
