+++ src/components/colors.js
