+++ src/components/helper-functions.js
@@ -1,0 +1,5 @@
+import { ToastAndroid } from 'react-native';
+
+export const toast = (message, length = ToastAndroid.SHORT) => {
+  ToastAndroid.show(message, length);
+};
