+++ src/screens/home-screen.js
@@ -1,27 +1,18 @@
-import {faFilter, faTimes} from '@fortawesome/free-solid-svg-icons';
-import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
-import {Container, Header, Icon, Input, Item} from 'native-base';
+import { faFilter, faTimes } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
+import { Container, Header, Icon, Input, Item } from 'native-base';
 import React from 'react';
-import {
-  FlatList,
-  Keyboard,
-  Modal,
-  SafeAreaView,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
-} from 'react-native';
+import { FlatList, Keyboard, Modal, SafeAreaView, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import Stars from 'react-native-stars';
 
 import Empty from '../assets/ratings/rating-empty-primary.png';
 import Full from '../assets/ratings/rating-full-primary.png';
-import {translate} from '../locales';
-import AbstractComponent from './abstract-component';
-import {getItem} from './common/async-storage-helper';
-import {toast} from './common/helper-functions';
-import LoadingSpinner from './common/loading-spinner';
-import MainCard from './common/main-card';
+import AbstractComponent from '../components/abstract-component';
+import { getItem } from '../components/async-storage';
+import { toast } from '../components/helper-functions';
+import LoadingSpinner from '../components/loading-spinner';
+import MainCard from '../components/main-card';
+import { translate } from '../locales';
 
 class Home extends AbstractComponent {
   constructor(props) {
